Designing a new GUI for the Server
