+++ src/server_chat/Server_Chat.java
@@ -1,15 +1,289 @@
 package server_chat;
 
+import java.awt.*;
+import java.io.*;
+import java.net.*;
+import java.util.*;
+import javax.swing.*;
 /**
  *
  * @author Alejandro
  */
+public class Server_Chat extends JFrame {
+    
+    private ObjectOutputStream output;
+    private ObjectInputStream input;
+    private ServerSocket server;
+    private Socket connection;
+    private LinkedList clientOutput;
+    private LinkedList clientList;
+    
+    public class ServerStart implements Runnable{
+        
+        @Override
+        public void run()
+        {
+            clientOutput = new LinkedList();
+            
+            try {
+            server = new ServerSocket(2222, 50);
 
-import javax.swing.*;
+                while(true){
+                    try{
+                        //Start listening
+                        ServerActionArea.append(" Waiting for someone to connect... \n");
+                        connection = server.accept();
+                        ServerActionArea.append(" Got a connection \n");
+                        //Create new client
+                        output = new ObjectOutputStream(connection.getOutputStream());
+                        clientOutput.add(output);
+                        output.flush();
+                        Thread client = new Thread(new ClientCreator(connection, output));
+                        client.start();
+                    }catch(EOFException eofException){
 
-public class Server_Chat {
+                    }
+                } 
+            }catch (IOException ioException){
 
-    public static void main(String[] args) {
-        Server_Setup chat = new Server_Setup();
+            }
+        }
     }
+    
+    public class ClientCreator implements Runnable {
+        
+        Socket client;
+        String username;
+
+        //Constructor
+        ClientCreator(Socket connection, ObjectOutputStream output) {
+            
+            client = connection;
+            
+            try {    
+                input = new ObjectInputStream(client.getInputStream());
+                String anon="anon";
+                Random generator = new Random(); 
+                int a = generator.nextInt(999);
+                String num = String.valueOf(a);
+                anon=anon.concat(num);
+                username=anon;
+            } catch (IOException ex) {
+                
+            }
+        }
+
+        @Override
+        public void run(){
+            
+            String message;
+            
+            try {                
+                while(((message = (String)input.readObject())) != null)
+                {
+                    sendMessage(message);
+                    if (message.equals(" USERNAME: is Disconnecting "))
+                    {
+                        clientOutput.remove(output);
+                    }
+                }
+            } catch (IOException | ClassNotFoundException ex) {
+                
+            }
+        }
+    }
+    
+    private void sendMessage(String message){
+        
+        Iterator it = clientOutput.iterator();
+        
+        while(it.hasNext()){
+            try{
+                ServerActionArea.append(" Sending: " + message + "\n");
+                ObjectOutputStream out = (ObjectOutputStream) it.next();
+                out.writeObject(message);
+                out.flush();
+            }catch(IOException ioException){
+
+            }            
+        }
+    }
+    
+    private void closeConnection(){
+        sendMessage(" The Server is shutting all connections... ");
+        ServerActionArea.append(" Closing connections... \n");
+        try{
+            clientOutput.removeAll(clientOutput);
+            output.close();
+            input.close();
+            connection.close();
+            server.close();
+        }catch(IOException ioException){
+            
+        }
+    }
+
+    public Server_Chat() {
+        initComponents();
+    }
+     
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        jPanel1 = new javax.swing.JPanel();
+        jScrollPane1 = new javax.swing.JScrollPane();
+        ServerActionArea = new javax.swing.JTextArea();
+        jScrollPane2 = new javax.swing.JScrollPane();
+        ServerUsersArea = new javax.swing.JTextArea();
+        ConnectionButton = new javax.swing.JButton();
+        UsersButton = new javax.swing.JButton();
+        jLabel1 = new javax.swing.JLabel();
+        jLabel2 = new javax.swing.JLabel();
+
+        javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
+        jPanel1.setLayout(jPanel1Layout);
+        jPanel1Layout.setHorizontalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 100, Short.MAX_VALUE)
+        );
+        jPanel1Layout.setVerticalGroup(
+            jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 100, Short.MAX_VALUE)
+        );
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+
+        ServerActionArea.setColumns(20);
+        ServerActionArea.setRows(5);
+        jScrollPane1.setViewportView(ServerActionArea);
+
+        ServerUsersArea.setColumns(20);
+        ServerUsersArea.setRows(5);
+        jScrollPane2.setViewportView(ServerUsersArea);
+
+        ConnectionButton.setText("Start Server");
+        ConnectionButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                ConnectionButtonActionPerformed(evt);
+            }
+        });
+
+        UsersButton.setText("Show Users");
+        UsersButton.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                UsersButtonActionPerformed(evt);
+            }
+        });
+
+        jLabel1.setText("Server Actions");
+
+        jLabel2.setText("Connected Users");
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 381, Short.MAX_VALUE)
+                    .addComponent(ConnectionButton, javax.swing.GroupLayout.PREFERRED_SIZE, 166, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addGap(18, 18, 18)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jScrollPane2)
+                    .addComponent(UsersButton, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                    .addComponent(jLabel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap())
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGap(6, 6, 6)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel1, javax.swing.GroupLayout.PREFERRED_SIZE, 19, javax.swing.GroupLayout.PREFERRED_SIZE)
+                    .addComponent(jLabel2))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                    .addComponent(jScrollPane1, javax.swing.GroupLayout.DEFAULT_SIZE, 399, Short.MAX_VALUE)
+                    .addComponent(jScrollPane2))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(ConnectionButton)
+                    .addComponent(UsersButton))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void ConnectionButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_ConnectionButtonActionPerformed
+        switch (ConnectionButton.getText()) {
+            case "Start Server":
+                Thread starter = new Thread(new ServerStart());
+                starter.start();
+                ConnectionButton.setText("End Server");
+                break;
+            case "End Server":
+                closeConnection();
+                ConnectionButton.setText("Start Server");
+                break;
+        }
+    }//GEN-LAST:event_ConnectionButtonActionPerformed
+
+    private void UsersButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_UsersButtonActionPerformed
+        Iterator it = clientOutput.iterator();
+        ServerUsersArea.append("\n Online users : \n");
+        while (it.hasNext()){
+            Object current_user = it.next();
+            ServerUsersArea.append((String) current_user);
+            ServerUsersArea.append("\n");
+        }
+    }//GEN-LAST:event_UsersButtonActionPerformed
+
+    public static void main(String args[]) {
+        /* Set the Nimbus look and feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
+         */
+        try {
+            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
+                if ("Nimbus".equals(info.getName())) {
+                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
+                    break;
+                }
+            }
+        } catch (ClassNotFoundException ex) {
+            java.util.logging.Logger.getLogger(Server_Chat.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (InstantiationException ex) {
+            java.util.logging.Logger.getLogger(Server_Chat.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (IllegalAccessException ex) {
+            java.util.logging.Logger.getLogger(Server_Chat.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
+            java.util.logging.Logger.getLogger(Server_Chat.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+        }
+        //</editor-fold>
+        //</editor-fold>
+
+        /* Create and display the form */
+        java.awt.EventQueue.invokeLater(new Runnable() {
+            public void run() {
+                new Server_Chat().setVisible(true);
+            }
+        });
+    }
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton ConnectionButton;
+    private javax.swing.JTextArea ServerActionArea;
+    private javax.swing.JTextArea ServerUsersArea;
+    private javax.swing.JButton UsersButton;
+    private javax.swing.JLabel jLabel1;
+    private javax.swing.JLabel jLabel2;
+    private javax.swing.JPanel jPanel1;
+    private javax.swing.JScrollPane jScrollPane1;
+    private javax.swing.JScrollPane jScrollPane2;
+    // End of variables declaration//GEN-END:variables
 }
